requirements.txt
streamlit
google-generativeai>=0.8.3
edge-tts